requirements.txt
tenacity
pydantic
supabase
functions-framework
python-dotenv
google-crc32c
google-cloud-secret-manager
psycopg2
pyyaml
git+https://github.com/Empyloo/supacrud.git
